fix: address Search API PR review feedback
Replace beforeEach setup with per-test factories per docs/testing.md and consolidate language-range parsing into schemas/search.ts.
